+++ hexen2/server/host_cmd.c
@@ -1,7 +1,7 @@
 /*
 	host_cmd.c
 
-	$Header: /home/ozzie/Download/0000/uhexen2/hexen2/server/host_cmd.c,v 1.4 2006-09-13 05:53:24 sezero Exp $
+	$Header: /home/ozzie/Download/0000/uhexen2/hexen2/server/host_cmd.c,v 1.5 2006-09-15 09:24:33 sezero Exp $
 */
 
 #include "quakedef.h"
@@ -230,7 +230,7 @@ static void Host_Map_f (void)
 		info_mask2 = 0;
 
 	svs.serverflags = 0;		// haven't completed an episode yet
-	strcpy (name, Cmd_Argv(1));
+	Q_strlcpy (name, Cmd_Argv(1), sizeof(name));
 
 	SV_SpawnServer (name, NULL);
 }
@@ -259,12 +259,12 @@ static void Host_Changelevel_f (void)
 		return;
 	}
 
-	strcpy (level, Cmd_Argv(1));
+	Q_strlcpy (level, Cmd_Argv(1), sizeof(level));
 	if (Cmd_Argc() == 2)
 		startspot = NULL;
 	else
 	{
-		strcpy (_startspot, Cmd_Argv(2));
+		Q_strlcpy (_startspot, Cmd_Argv(2), sizeof(_startspot));
 		startspot = _startspot;
 	}
 
@@ -293,8 +293,8 @@ static void Host_Restart_f (void)
 		return;
 	}
 
-	strcpy (mapname, sv.name);	// must copy out, because it gets cleared
-	strcpy(startspot, sv.startspot);
+	Q_strlcpy (mapname, sv.name, sizeof(mapname));	// must copy out, because it gets cleared
+	Q_strlcpy (startspot, sv.startspot, sizeof(startspot));
 
 	if (Cmd_Argc() == 2 && Q_strcasecmp(Cmd_Argv(1),"restore") == 0)
 	{
@@ -423,22 +423,34 @@ static void Host_Savegame_f (void)
 	if (error_state)
 		return;
 
-	sprintf (savename, "%s/%s", com_savedir, Cmd_Argv(1));
-	Sys_mkdir (savename);
+	if (snprintf(savename, sizeof(savename), "%s/%s", com_savedir, Cmd_Argv(1)) >= sizeof(savename))
+	{
+		Con_Printf ("%s: save directory name too long\n", __FUNCTION__);
+		return;
+	}
+	if (Sys_mkdir(savename))
+	{
+		Con_Printf ("Unable to create save directory\n");
+		return;
+	}
 
 	Host_RemoveGIPFiles(savename);
 
-	sprintf (savename, "%s/clients.gip", com_savedir);
+	snprintf (savename, sizeof(savename), "%s/clients.gip", com_savedir);
 	unlink(savename);
 
-	sprintf (savedest, "%s/%s",com_savedir, Cmd_Argv(1));
+	snprintf (savedest, sizeof(savedest), "%s/%s", com_savedir, Cmd_Argv(1));
 	Con_Printf ("Saving game to %s...\n", savedest);
 
 	error_state = Host_CopyFiles(com_savedir, "*.gip", savedest);
 	if (error_state)
 		goto retrymsg;
 
-	sprintf(savedest,"%s/%s/info.dat",com_savedir, Cmd_Argv(1));
+	if (snprintf(savedest, sizeof(savedest), "%s/%s/info.dat", com_savedir, Cmd_Argv(1)) >= sizeof(savedest))
+	{
+		Con_Printf ("%s: string buffer overflow!\n", __FUNCTION__);
+		return;
+	}
 	f = fopen (savedest, "w");
 	if (!f)
 	{
@@ -505,16 +517,23 @@ static void Host_Loadgame_f (void)
 
 	Host_RemoveGIPFiles(NULL);
 
-	sprintf (savename, "%s/%s", com_savedir, Cmd_Argv(1));
-
+	if (snprintf(savename, sizeof(savename), "%s/%s", com_savedir, Cmd_Argv(1)) >= sizeof(savename))
+	{
+		Con_Printf ("%s: save directory name too long\n", __FUNCTION__);
+		return;
+	}
 	Con_Printf ("Loading game from %s...\n", savename);
 
-	sprintf(savedest,"%s/info.dat",savename);
+	if (snprintf(savedest, sizeof(savedest), "%s/info.dat", savename) >= sizeof(savedest))
+	{
+		Con_Printf ("%s: string buffer overflow!\n", __FUNCTION__);
+		return;
+	}
 
 	f = fopen (savedest, "r");
 	if (!f)
 	{
-		Con_Printf ("ERROR: couldn't open.\n");
+		Con_Printf ("%s: ERROR: couldn't open savefile\n", __FUNCTION__);
 		return;
 	}
 
@@ -580,8 +599,7 @@ static void Host_Loadgame_f (void)
 
 	Host_RemoveGIPFiles(com_savedir);
 
-	sprintf (savedest, "%s/%s",com_savedir, Cmd_Argv(1));
-
+	snprintf (savedest, sizeof(savedest), "%s/%s", com_savedir, Cmd_Argv(1));
 	error_state = Host_CopyFiles(savedest, "*.gip", com_savedir);
 
 	if (error_state)
@@ -622,15 +640,22 @@ qboolean SaveGamestate(qboolean ClientsOnly)
 		start = 1;
 		end = svs.maxclients+1;
 
-		sprintf (savename, "%s/clients.gip", com_savedir);
+		if (snprintf(savename, sizeof(savename), "%s/clients.gip", com_savedir) >= sizeof(savename))
+		{
+			Con_Printf ("%s: string buffer overflow!\n", __FUNCTION__);
+			return true;
+		}
 	}
 	else
 	{
 		start = 1;
 		end = sv.num_edicts;
 
-		sprintf (savename, "%s/%s.gip", com_savedir, sv.name);
-//		Con_Printf ("Saving game to %s...\n", savename);
+		if (snprintf(savename, sizeof(savename), "%s/%s.gip", com_savedir, sv.name) >= sizeof(savename))
+		{
+			Con_Printf ("%s: string buffer overflow!\n", __FUNCTION__);
+			return true;
+		}
 	}
 
 	f = fopen (savename, "w");
@@ -782,11 +807,19 @@ static int LoadGamestate(char *level, char *startspot, int ClientsMode)
 
 	if (ClientsMode == 1)
 	{
-		sprintf (savename, "%s/clients.gip", com_savedir);
+		if (snprintf(savename, sizeof(savename), "%s/clients.gip", com_savedir) >= sizeof(savename))
+		{
+			Con_Printf ("%s: string buffer overflow!\n", __FUNCTION__);
+			return -1;
+		}
 	}
 	else
 	{
-		sprintf (savename, "%s/%s.gip", com_savedir, level);
+		if (snprintf(savename, sizeof(savename), "%s/%s.gip", com_savedir, level) >= sizeof(savename))
+		{
+			Con_Printf ("%s: string buffer overflow!\n", __FUNCTION__);
+			return -1;
+		}
 
 		if (ClientsMode != 2 && ClientsMode != 3)
 			Con_Printf ("Loading game from %s...\n", savename);
@@ -796,7 +829,7 @@ static int LoadGamestate(char *level, char *startspot, int ClientsMode)
 	if (!f)
 	{
 		if (ClientsMode == 2)
-			Con_Printf ("ERROR: couldn't open.\n");
+			Con_Printf ("%s: ERROR: couldn't open savefile\n", __FUNCTION__);
 
 		return -1;
 	}
@@ -967,12 +1000,12 @@ static void Host_Changelevel2_f (void)
 		return;
 	}
 
-	strcpy (level, Cmd_Argv(1));
+	Q_strlcpy (level, Cmd_Argv(1), sizeof(level));
 	if (Cmd_Argc() == 2)
 		startspot = NULL;
 	else
 	{
-		strcpy (_startspot, Cmd_Argv(2));
+		Q_strlcpy (_startspot, Cmd_Argv(2), sizeof(_startspot));
 		startspot = _startspot;
 	}
 
@@ -1010,7 +1043,7 @@ static void Host_Name_f (void)
 		newName = Cmd_Argv(1);
 	else
 		newName = Cmd_Args();
-	newName[15] = 0;
+	newName[15] = 0;	// client_t structure actually says name[32].
 
 	//this is for the fuckers who put braces in the name causing loadgame to crash.
 	pdest = strchr(newName,'{');
@@ -1173,16 +1206,13 @@ static void Host_Say(qboolean teamonly)
 
 // turn on color set 1
 	if (!fromServer)
-		sprintf (text, "%c%s: ", 1, save->name);
+		snprintf (text, sizeof(text), "%c%s: ", 1, save->name);
 	else
-		sprintf (text, "%c<%s> ", 1, hostname.string);
+		snprintf (text, sizeof(text), "%c<%s> ", 1, hostname.string);
 
-	j = sizeof(text) - 2 - strlen(text);  // -2 for /n and null terminator
-	if (strlen(p) > j)
-		p[j] = 0;
-
-	strcat (text, p);
-	strcat (text, "\n");
+	Q_strlcat (text, p, sizeof(text));
+	if (Q_strlcat (text, "\n", sizeof(text)) >= sizeof(text))
+		text[sizeof(text)-2] = '\n';
 
 	for (j = 0, client = svs.clients; j < svs.maxclients; j++, client++)
 	{
@@ -1225,8 +1255,8 @@ static void Host_Tell_f(void)
 	if (Cmd_Argc () < 3)
 		return;
 
-	strcpy(text, host_client->name);
-	strcat(text, ": ");
+	Q_strlcpy(text, host_client->name, sizeof(text));
+	Q_strlcat(text, ": ", sizeof(text));
 
 	p = Cmd_Args();
 
@@ -1238,12 +1268,9 @@ static void Host_Tell_f(void)
 	}
 
 // check length & truncate if necessary
-	j = sizeof(text) - 2 - strlen(text);  // -2 for /n and null terminator
-	if (strlen(p) > j)
-		p[j] = 0;
-
-	strcat (text, p);
-	strcat (text, "\n");
+	Q_strlcat (text, p, sizeof(text));
+	if (Q_strlcat (text, "\n", sizeof(text)) >= sizeof(text))
+		text[sizeof(text)-2] = '\n';
 
 	save = host_client;
 	for (j = 0, client = svs.clients; j < svs.maxclients; j++, client++)
@@ -1730,6 +1757,10 @@ void Host_InitCommands (void)
 
 /*
  * $Log: not supported by cvs2svn $
+ * Revision 1.4  2006/09/13 05:53:24  sezero
+ * re-visited the includes, gathered all net includes into
+ * the new net_sys.h, did a platform defines clean-up.
+ *
  * Revision 1.3  2006/07/02 11:36:36  sezero
  * uppercased the pr_global_struct() macro for easier detection
  * and searching. put that macro in use in hexenworld server for
